requirement.txt
pandas
tqdm
fuzzywuzzy
requests
imagehash
Pillow
matplotlib
